# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Recovery Velocity Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a short-lived upload URL for citizen-report photo evidence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Unsupported image metadata
        "429":
          description: Too many upload requests
...
  /tasks/dispatch:
    post:
      operationId: dispatchTask
      tags: [tasks]
      summary: Coordinator dispatches an approved incident to a field officer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskDispatchInput"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Validation error
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks, optionally filtered by officer
      parameters:
        - in: query
          name: officer
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Task list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
  /tasks/{id}/complete:
    post:
      operationId: completeTask
      tags: [tasks]
      summary: Backward-compatible field submission that moves a task to pending verification
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCompleteInput"
      responses:
        "200":
          description: Task completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Task not found
  /tasks/{id}/start:
    post:
      operationId: startTask
      tags: [tasks]
      summary: Assigned field unit leader starts their task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskActorInput"
      responses:
        "200":
          description: Task started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "409":
          description: Invalid transition or actor
  /tasks/{id}/submit:
    post:
      operationId: submitTask
      tags: [tasks]
      summary: Assigned field unit leader submits evidence for OPS verification
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskSubmissionInput"
      responses:
        "200":
          description: Task submitted for verification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "409":
          description: Invalid transition or actor
  /tasks/{id}/verify:
    post:
      operationId: verifyTask
      tags: [tasks]
      summary: OPS Section Chief verifies submitted evidence and completes the task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskVerificationInput"
      responses:
        "200":
          description: Task verified and completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "409":
          description: Invalid transition or actor
...
components:
  schemas:
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        size:
          type: number
          minimum: 1
          maximum: 5242880
        contentType:
          type: string
          enum: [image/jpeg, image/png, image/webp]
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath, metadata]
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"
    Task:
      type: object
      required: [id, incidentId, title, parishId, parishName, officer, instructions, status, createdAt, statusChangedAt, scoreImpact]
      properties:
        id:
          type: string
        incidentId:
          type: number
        title:
          type: string
        parishId:
          type: string
        parishName:
          type: string
        officer:
          type: string
        instructions:
          type: string
        scoreArea:
          type: string
          enum: [readiness, recovery, resilience, incident_risk]
        recommendationId:
          type: string
        expectedImpact:
          type: string
        scoreImpact:
          $ref: "#/components/schemas/TaskScoreImpact"
        evidenceStatus:
          type: string
          enum: [planning_guidance, partner_reported, evidence_backed]
        status:
          type: string
          enum: [assigned, in_progress, pending_verification, completed]
        createdAt:
          type: string
        statusChangedAt:
          type: string
        startedAt:
          type: string
        submittedAt:
          type: string
        completedAt:
          type: string
        fieldNote:
          type: string
        verifiedAt:
          type: string
        verifiedBy:
          type: string
        lastActor:
          type: string
    TaskDispatchInput:
      type: object
      required: [incidentId, title, parishId, parishName, officer, instructions]
      properties:
        incidentId:
          type: number
        title:
          type: string
        parishId:
          type: string
        parishName:
          type: string
        officer:
          type: string
        instructions:
          type: string
        scoreArea:
          type: string
          enum: [readiness, recovery, resilience, incident_risk]
        recommendationId:
          type: string
        expectedImpact:
          type: string
        impactPoints:
          type: number
        evidenceStatus:
          type: string
          enum: [planning_guidance, partner_reported, evidence_backed]
    TaskCompleteInput:
      type: object
      required: [fieldNote]
      properties:
        fieldNote:
          type: string
          minLength: 1
        actor:
          type: string
    TaskActorInput:
      type: object
      required: [actor]
      properties:
        actor:
          type: string
          minLength: 1
    TaskSubmissionInput:
      type: object
      required: [actor, fieldNote]
      properties:
        actor:
          type: string
          minLength: 1
        fieldNote:
          type: string
          minLength: 1
    TaskVerificationInput:
      type: object
      required: [actor]
      properties:
        actor:
          type: string
          minLength: 1
    TaskScoreImpact:
      type: object
      required: [area, points, unlocked]
      properties:
        area:
          type: string
          enum: [readiness, recovery, resilience, incident_risk]
        points:
          type: number
        unlocked:
          type: boolean
...